requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
openai>=1.30.0
qdrant-client>=1.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.33.0
pyperclip>=1.8.2
python-docx>=1.1.0
